feat: RestDocs 설정 #241
- default snippets 중에서 `curl-request`와 `httpie-request`를 사용하지 않아 수정했습니다.
- request, response에 pretty print 옵션을 넣어서 json 형식을 출력합니다.
- util 패키지에 IntegrationTest를 넣었습니다.
