--- be/src/test/java/codesquad/bookkbookk/IntegrationTest.java
@@ -1,34 +1,0 @@
-package codesquad.bookkbookk;
-
-import org.junit.jupiter.api.BeforeEach;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.web.server.LocalServerPort;
-import org.springframework.data.redis.core.RedisTemplate;
-import org.springframework.test.context.ActiveProfiles;
-
-import codesquad.bookkbookk.util.DatabaseCleaner;
-
-import io.restassured.RestAssured;
-
-@ActiveProfiles(profiles = "test")
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-public class IntegrationTest {
-
-    @LocalServerPort
-    private int port;
-
-    @Autowired
-    private DatabaseCleaner databaseCleaner;
-
-    @Autowired
-    private RedisTemplate<String, Object> redisTemplate;
-
-    @BeforeEach
-    public void setUp() {
-        databaseCleaner.execute();
-        redisTemplate.getConnectionFactory().getConnection().flushAll();
-        RestAssured.port = port;
-    }
-
-}
