+++ be/src/test/java/codesquad/bookkbookk/util/IntegrationTest.java
@@ -1,0 +1,57 @@
+package codesquad.bookkbookk.util;
+
+import static org.springframework.restdocs.http.HttpDocumentation.*;
+import static org.springframework.restdocs.operation.preprocess.Preprocessors.*;
+import static org.springframework.restdocs.payload.PayloadDocumentation.*;
+import static org.springframework.restdocs.restassured3.RestAssuredRestDocumentation.*;
+
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.test.web.server.LocalServerPort;
+import org.springframework.data.redis.core.RedisTemplate;
+import org.springframework.restdocs.RestDocumentationContextProvider;
+import org.springframework.restdocs.RestDocumentationExtension;
+import org.springframework.test.context.ActiveProfiles;
+
+import io.restassured.RestAssured;
+import io.restassured.builder.RequestSpecBuilder;
+import io.restassured.specification.RequestSpecification;
+
+@ActiveProfiles(profiles = "test")
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
+@ExtendWith(RestDocumentationExtension.class)
+public class IntegrationTest {
+
+    @LocalServerPort
+    private int port;
+
+    @Autowired
+    private DatabaseCleaner databaseCleaner;
+
+    @Autowired
+    private RedisTemplate<String, Object> redisTemplate;
+
+    protected RequestSpecification spec;
+
+    @BeforeEach
+    public void setUp(RestDocumentationContextProvider restDocumentation) {
+        this.databaseCleaner.execute();
+        this.redisTemplate.getConnectionFactory().getConnection().flushAll();
+        RestAssured.port = port;
+        this.spec = new RequestSpecBuilder()
+                .addFilter(documentationConfiguration(restDocumentation)
+                        .snippets().withDefaults(httpRequest(), httpResponse(), requestBody(), responseBody())
+                        .and()
+                        .operationPreprocessors()
+                        .withRequestDefaults(modifyUris()
+                                .scheme("https")
+                                .host("bookkbookk.site")
+                                .removePort(),
+                                prettyPrint())
+                        .withResponseDefaults(prettyPrint()))
+                .build();
+    }
+
+}
